+++ ngClient/App/settings/Setup/googleAnalytics.js
@@ -2,7 +2,7 @@
     'use strict';
 
     angular.module('main')
-        .constant('AnalyticsTrackingCode', '') // Enter your google analytics tracking code here
-        .constant('AnalyticsDomainName', 'none'); // Enter your domain name here - 'none' for localhost
+        .constant('analyticsTrackingCode', '') // Enter your google analytics tracking code here
+        .constant('analyticsDomainName', 'none'); // Enter your domain name here - 'none' for localhost
 
 })();